flask/requirements.txt
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.13
attrs==23.1.0
autopep8==2.0.0
cffi==1.15.1
click==8.1.3
colorlog==6.7.0
dill==0.3.6
flake8==6.0.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
flask-restx==1.1.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
marshmallow==3.19.0
mccabe==0.7.0
packaging==23.1
peewee==3.16.2
platformdirs==2.6.0
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==0.21.0
pytz==2023.3
tomli==2.0.1
tomlkit==0.11.6
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
